liblubi: Fix off-by-one usage of max_lnum in read_svol
In this change, the offset vs CFG_LUBI_PEB_NB_MAX is not important,
however the area of leb2pebs which is memset is.
The consequence is a possible bad mapping of the last LEB (i.e with
highest lnum) of a volume.

I just found out this bug while integrating liblubi into the GRX SPL.

Signed-off-by: Karl Beldan <karl.beldan-ext@sagemcom.com>